test swap and fmt

+++ foundry/test/uniswap-v2/exercises/UniswapV2Liquidity.test.sol
@@ -46,19 +46,20 @@ contract UniswapV2LiquidityTest is Test {
         // Write your code here
         // Don’t change any other code
         vm.prank(user);
-        (uint amountA, uint amountB, uint liquidity) = router.addLiquidity({
+        (uint256 amountA, uint256 amountB, uint256 liquidity) = router
+            .addLiquidity({
             tokenA: DAI,
             tokenB: WETH,
             amountADesired: 1e6 * 1e18,
-            amountBDesired:100 * 1e18,
-            amountAMin:1,
-            amountBMin:1,
-            to:user,
-            deadline:block.timestamp
+            amountBDesired: 100 * 1e18,
+            amountAMin: 1,
+            amountBMin: 1,
+            to: user,
+            deadline: block.timestamp
         });
-        console2.log('DAI', amountA);
-        console2.log('WETH', amountB);
-        console2.log('Liquidity', liquidity);
+        console2.log("DAI", amountA);
+        console2.log("WETH", amountB);
+        console2.log("Liquidity", liquidity);
         assertGt(pair.balanceOf(user), 0, "LP = 0");
     }
 
@@ -79,18 +80,18 @@ contract UniswapV2LiquidityTest is Test {
         // Exercise - Remove liquidity from DAI / WETH pool
         // Write your code here
         // Don’t change any other code
-        (uint amountA, uint amountB) = router.removeLiquidity({
-        tokenA: WETH,
-        tokenB: DAI,
-        liquidity: liquidity,
-        amountAMin: 1,
-        amountBMin: 1,
-        to: user,
-        deadline: block.timestamp
-        }); 
+        (uint256 amountA, uint256 amountB) = router.removeLiquidity({
+            tokenA: WETH,
+            tokenB: DAI,
+            liquidity: liquidity,
+            amountAMin: 1,
+            amountBMin: 1,
+            to: user,
+            deadline: block.timestamp
+        });
         vm.stopPrank();
-        console2.log('DAI', amountA);
-        console2.log('WETH', amountB);
+        console2.log("DAI", amountA);
+        console2.log("WETH", amountB);
 
         assertEq(pair.balanceOf(user), 0, "LP = 0");
     }
